+++ contracts/core/accounts/interface/IAccountsCreateEndowment.sol
@@ -6,5 +6,5 @@ import {AccountMessages} from "../message.sol";
 interface IAccountsCreateEndowment {
     function createEndowment(
         AccountMessages.CreateEndowmentRequest memory curDetails
-    ) external returns (uint256);
+    ) external returns (uint32);
 }
