+++ typechain-types/factories/contracts/normalized_endowment/subdao/ISubdaoEmitter__factory.ts
@@ -20,9 +20,9 @@ const _abi = [
       {
         components: [
           {
-            internalType: "uint256",
+            internalType: "uint32",
             name: "id",
-            type: "uint256",
+            type: "uint32",
           },
           {
             internalType: "address",
